Migrate Logic into organized files
